final_project/API-Chalice/requirements.txt
chalice
psycopg2-binary
